fix(coverage): add 'infrastructure' hint so monthly §3 boost fires
Reviewer finding: _SECTION_CATEGORY_HINTS missed 'infrastructure', so
_section_required_category returned None for the monthly brief's
'Frontier & Infrastructure' title. The new 3.0 per-category boost in
_score_source was a no-op on the brief that motivated the change.

Add 'infrastructure' -> 'official' (one line). Also tighten
_score_source's return annotation from int to float — pre-existing
hybrid made worse by the new 3.0 literal.

Regression test pins both 'Frontier & Infrastructure' and
'Hardware & Infrastructure' so the map can't regress silently.

+++ src/hermes/pipeline/synthesize.py
@@ -51,7 +51,7 @@ def _recency_bonus(s: SearchResult, recency_days: int | None) -> int:
         return 0
 
 
-def _score_source(kws: set[str], s: SearchResult, recency_days: int | None, *, section: SectionSpec | None = None) -> int:
+def _score_source(kws: set[str], s: SearchResult, recency_days: int | None, *, section: SectionSpec | None = None) -> float:
     title = (s.title or "").lower()
     body = f"{s.title} {s.content}".lower()
     score = 0
